refactor(reflex4you): share constant evaluator
